requirements.txt
# Python dependencies for RAG functionality
# PDF processing with Docling
docling>=2.7.0
transformers>=4.50.3
# Groq API client for LLM generation
groq>=0.11.0
# HTTP client for Jina API
httpx>=0.28.0
# Weaviate client for vector search
weaviate-client>=4.0.0
# Pydantic for input validation
pydantic>=2.0.0
